+++ my-game/Game.h
@@ -67,7 +67,7 @@ public:
 	void InitKeyboard();
 	void SetKeyHandler(LPKEYEVENTHANDLER handler) { keyHandler = handler; }
 	void Init(HWND hWnd);
-	void Draw(float x, float y, LPDIRECT3DTEXTURE9 texture, int left, int top, int right, int bottom, int alpha = 255);
+	void Draw(float x, float y, LPDIRECT3DTEXTURE9 texture, float left, float top, float right, float bottom, int alpha = 255);
 	void DrawWithScale(Vector p, LPDIRECT3DTEXTURE9 texture, Rect r, int opacity, D3DXVECTOR2 scale);
 	void DrawPositionInCamera(Vector p, LPDIRECT3DTEXTURE9 texture, Rect r, int opacity, D3DXVECTOR2 scale);
